requirements.txt
astroid==3.3.11
dans-toolbox @ file:///home/dan/Desktop/Toolbox
dill==0.4.0
iniconfig==2.1.0
isort==6.0.1
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mccabe==0.7.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.3.3
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
Pygments==2.19.2
pyright==1.1.405
pytest==8.4.2
rich==14.1.0
textual==6.1.0
textual-imageview==0.1.1
tomlkit==0.13.3
toolbox==1.11.0
typing_extensions==4.15.0
uc-micro-py==1.0.3
